Harden SDK conversation adapters
