+++ scripts/check-tank-conversation-contract.mjs
@@ -7,17 +7,23 @@ import { createRequire } from "node:module";
 
 const repoRoot = path.resolve(path.dirname(fileURLToPath(import.meta.url)), "..");
 const schemaPath = path.join(repoRoot, "schemas", "tank-conversation-event.schema.json");
+const fixturePath = path.join(repoRoot, "schemas", "tank-conversation-event.fixtures.json");
 const requireFromFrontend = createRequire(path.join(repoRoot, "frontend", "package.json"));
 let ts;
+let Ajv2020;
+let addFormats;
 try {
   ts = requireFromFrontend("typescript");
+  Ajv2020 = requireDefault(requireFromFrontend("ajv/dist/2020"));
+  addFormats = requireDefault(requireFromFrontend("ajv-formats"));
 } catch (err) {
-  console.error("Unable to load TypeScript from frontend/node_modules.");
+  console.error("Unable to load contract check dependencies from frontend/node_modules.");
   console.error("Run `npm ci --prefix frontend` before this contract check.");
   throw err;
 }
 
 const schema = JSON.parse(await fs.readFile(schemaPath, "utf8"));
+const fixtures = JSON.parse(await fs.readFile(fixturePath, "utf8"));
 const expectedEnums = {
   TANK_ACTORS: schemaEnum("actor"),
   TANK_EVENT_SOURCES: schemaEnum("source"),
@@ -47,13 +53,21 @@ for (const relativePath of tsContractFiles) {
   }
 }
 
+const validatedFixtureCount = validateFixtures();
+
 if (failures.length > 0) {
   console.error("Tank conversation contract drift detected:");
   for (const failure of failures) console.error(`- ${failure}`);
   process.exit(1);
 }
 
-console.log("Tank conversation TypeScript enums match schemas/tank-conversation-event.schema.json");
+console.log(
+  `Tank conversation contract matches schemas/tank-conversation-event.schema.json; validated ${validatedFixtureCount} canonical fixtures.`,
+);
+
+function requireDefault(module) {
+  return module.default ?? module;
+}
 
 function schemaEnum(propertyName) {
   const values = schema?.properties?.[propertyName]?.enum;
@@ -78,6 +92,95 @@ function compareArrays(label, actual, expected) {
   if (actual.length === expected.length && actual.some((value, index) => value !== expected[index])) {
     failures.push(`${label}: values match but order differs from the schema`);
   }
+}
+
+function validateFixtures() {
+  const fixtureEvents = fixtures?.events;
+  if (!Array.isArray(fixtureEvents)) {
+    failures.push("schemas/tank-conversation-event.fixtures.json: missing events array");
+    return 0;
+  }
+
+  const ajv = new Ajv2020({ allErrors: true, strict: true, strictRequired: false });
+  addFormats(ajv);
+  const validate = ajv.compile(schema);
+  const fixtureTypes = new Set();
+  let sawRunnerStampedEvent = false;
+  let sawPersistedEvent = false;
+
+  for (const [index, fixture] of fixtureEvents.entries()) {
+    const label = fixtureLabel(fixture, index);
+    const event = fixture?.event;
+    if (!event || typeof event !== "object" || Array.isArray(event)) {
+      failures.push(`${label}: missing event object`);
+      continue;
+    }
+    if (typeof event.type === "string") fixtureTypes.add(event.type);
+    if (hasRunnerStamps(event)) sawRunnerStampedEvent = true;
+    if (hasStorageStamps(event)) sawPersistedEvent = true;
+    if (!validate(event)) {
+      failures.push(`${label}: ${formatAjvErrors(validate.errors)}`);
+    }
+  }
+
+  for (const eventType of expectedEnums.TANK_EVENT_TYPES) {
+    if (!fixtureTypes.has(eventType)) {
+      failures.push(
+        `schemas/tank-conversation-event.fixtures.json: missing fixture for ${JSON.stringify(eventType)}`,
+      );
+    }
+  }
+  for (const eventType of fixtureTypes) {
+    if (!expectedEnums.TANK_EVENT_TYPES.includes(eventType)) {
+      failures.push(
+        `schemas/tank-conversation-event.fixtures.json: fixture uses unknown type ${JSON.stringify(eventType)}`,
+      );
+    }
+  }
+  if (!sawRunnerStampedEvent) {
+    failures.push("schemas/tank-conversation-event.fixtures.json: expected at least one runner-stamped fixture");
+  }
+  if (!sawPersistedEvent) {
+    failures.push("schemas/tank-conversation-event.fixtures.json: expected at least one persisted timeline fixture");
+  }
+
+  return fixtureEvents.length;
+}
+
+function fixtureLabel(fixture, index) {
+  const name = typeof fixture?.name === "string" && fixture.name ? fixture.name : `fixture ${index + 1}`;
+  return `schemas/tank-conversation-event.fixtures.json:${name}`;
+}
+
+function hasRunnerStamps(event) {
+  return (
+    typeof event.uuid === "string" &&
+    Number.isInteger(event.tank_event_seq) &&
+    typeof event.tank_order_key === "string" &&
+    typeof event.written_at === "string"
+  );
+}
+
+function hasStorageStamps(event) {
+  return (
+    typeof event.id === "string" &&
+    typeof event.tank_session_id === "string" &&
+    typeof event.email === "string" &&
+    typeof event.runtime === "string"
+  );
+}
+
+function formatAjvErrors(errors) {
+  if (!errors || errors.length === 0) return "schema validation failed";
+  return errors
+    .map((error) => {
+      const location = error.instancePath || "/";
+      const allowed = Array.isArray(error.params?.allowedValues)
+        ? `: ${error.params.allowedValues.map((value) => JSON.stringify(value)).join(", ")}`
+        : "";
+      return `${location} ${error.message}${allowed}`;
+    })
+    .join("; ");
 }
 
 function stringArrayConst(sourceFile, constName) {
